introduce an option to add a trailing fat partition
So far this only handles dos partition tables.
When ADD_DOS_PART is set to 1, add a small 4MB fat filesystem as the
last partition to this image, and copy the two DietPi config files to
this partition.
Now even people who don't have access to a Linux system can very easily
mount the DIETPISETUP partition, edit the two files and get the setup
they want. This is especially important for headless devices connecte
via WiFi - where this initial setup is necessary for them to connect to
the network in the first place.

This commit only creates that new partition. The setup logic still needs
to be updated so that it mounts that partition (when present) and uses
the config files on that partition instead of the ones provided
elsewhere in the image.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>